server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.5.0
python-dotenv==1.0.0
httpx==0.27.0
pytest==7.4.3
pytest-asyncio==0.21.1
email-validator==2.1.0
